Initial (v1.0.0)
Adds a GUI program, and a console program

You can make it so the only option is the key but the console version already has that for you newbie/non-coders

The GUI edition lets you generate up to 9999 keys, Please update your configs!

The console edition is very minimal and only lets you enter keys.